Add conditional removal of images

+++ app.py
@@ -27,15 +27,17 @@ def upload():
     # Deletes any file in the user-uploads
     folder = './static/user-uploads'
 
-    for filename in os.listdir(folder):
-        file_path = os.path.join(folder, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+    # Remove all files in user-uploads if there are any
+    if len(os.listdir(folder)) != 0:
+        for filename in os.listdir(folder):
+            file_path = os.path.join(folder, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
 
     # The file name is the time of upload
     upload_time = dt.now().strftime('%Y%m%d%H%M%S%f')
